refactor: isolate request runtime policy
